Continued splitting code into logical files
Fixed CommandMessage Bug as well
