+++ hardware/plugins/Plugins.cpp
@@ -14,11 +14,10 @@
 #include "../main/Logger.h"
 #include "../main/SQLHelper.h"
 #include "../main/mainworker.h"
-#include "../hardware/hardwaretypes.h"
 #include "../tinyxpath/tinyxml.h"
 #include "../main/localtime_r.h"
 
-#include "DelayedLink.h"
+#include "PythonObjects.h"
 
 #define ADD_STRING_TO_DICT(pDict, key, value) \
 		{	\
@@ -499,7 +498,7 @@ namespace Plugins {
 		return 0;
 	}
 
-	static struct PyModuleDef DomoticzModuleDef = {
+	struct PyModuleDef DomoticzModuleDef = {
 		PyModuleDef_HEAD_INIT,
 		"Domoticz",
 		NULL,
@@ -509,594 +508,6 @@ namespace Plugins {
 		DomoticzTraverse,
 		DomoticzClear,
 		NULL
-	};
-
-	typedef struct {
-		PyObject_HEAD
-		PyObject*	PluginKey;
-		int			HwdID;
-		PyObject*	DeviceID;
-		int			Unit;
-		int			Type;
-		int			SubType;
-		int			SwitchType;
-		int			ID;
-		int			LastLevel;
-		PyObject*	Name;
-		PyObject*	LastUpdate;
-		int			nValue;
-		int			SignalLevel;
-		int			BatteryLevel;
-		PyObject*	sValue;
-		int			Image;
-		PyObject*	Options;
-		CPlugin*	pPlugin;
-	} CDevice;
-
-	static void CDevice_dealloc(CDevice* self)
-	{
-		Py_XDECREF(self->Name);
-		Py_XDECREF(self->sValue);
-		Py_TYPE(self)->tp_free((PyObject*)self);
-	}
-
-	static PyObject* CDevice_new(PyTypeObject *type, PyObject *args, PyObject *kwds)
-	{
-		CDevice *self = (CDevice *)type->tp_alloc(type, 0);
-
-		try
-		{
-			if (self == NULL) {
-				_log.Log(LOG_ERROR, "%s: Self is NULL.", __func__);
-			}
-			else {
-				self->PluginKey = PyUnicode_FromString("");
-				if (self->PluginKey == NULL) {
-					Py_DECREF(self);
-					return NULL;
-				}
-				self->HwdID = -1;
-				self->DeviceID = PyUnicode_FromString("");
-				if (self->DeviceID == NULL) {
-					Py_DECREF(self);
-					return NULL;
-				}
-				self->Unit = -1;
-				self->Type = 0;
-				self->SubType = 0;
-				self->SwitchType = 0;
-				self->ID = -1;
-				self->LastLevel;
-				self->Name = PyUnicode_FromString("");
-				if (self->Name == NULL) {
-					Py_DECREF(self);
-					return NULL;
-				}
-				self->nValue = 0;
-				self->sValue = PyUnicode_FromString("");
-				if (self->sValue == NULL) {
-					Py_DECREF(self);
-					return NULL;
-				}
-				self->Options = PyUnicode_FromString("");
-				if (self->Options == NULL) {
-					Py_DECREF(self);
-					return NULL;
-				}
-				self->Image = 0;
-				self->SignalLevel = 100;
-				self->BatteryLevel = 255;
-				self->pPlugin = NULL;
-			}
-		}
-		catch (std::exception e)
-		{
-			_log.Log(LOG_ERROR, "%s: Execption thrown: %s", __func__, e.what());
-		}
-		catch (...)
-		{
-			_log.Log(LOG_ERROR, "%s: Unknown execption thrown", __func__);
-		}
-
-		return (PyObject *)self;
-	}
-
-	static int CDevice_init(CDevice *self, PyObject *args, PyObject *kwds)
-	{
-		char*		Name = NULL;
-		int			Unit = -1;
-		char*		TypeName = NULL;
-		int			Type = -1;
-		int			SubType = -1;
-		int			SwitchType = -1;
-		int			Image = -1;
-		char*		Options = NULL;
-		static char *kwlist[] = { "Name", "Unit", "TypeName", "Type", "Subtype", "Switchtype", "Image", "Options", NULL };
-
-		try
-		{
-			PyObject*	pModule = PyState_FindModule(&DomoticzModuleDef);
-			if (!pModule)
-			{
-				_log.Log(LOG_ERROR, "CPlugin:%s, unable to find module for current interpreter.", __func__);
-				return 0;
-			}
-
-			module_state*	pModState = ((struct module_state*)PyModule_GetState(pModule));
-			if (!pModState)
-			{
-				_log.Log(LOG_ERROR, "CPlugin:%s, unable to obtain module state.", __func__);
-				return 0;
-			}
-
-			if (!pModState->pPlugin)
-			{
-				_log.Log(LOG_ERROR, "CPlugin:%s, illegal operation, Plugin has not started yet.", __func__);
-				return 0;
-			}
-
-			if (PyArg_ParseTupleAndKeywords(args, kwds, "si|siiiis", kwlist, &Name, &Unit, &TypeName, &Type, &SubType, &SwitchType, &Image, &Options))
-			{
-				self->pPlugin = pModState->pPlugin;
-				self->PluginKey = PyUnicode_FromString(pModState->pPlugin->m_PluginKey.c_str());
-				self->HwdID = pModState->pPlugin->m_HwdID;
-				if (Name) {
-					Py_DECREF(self->Name);
-					self->Name = PyUnicode_FromString(Name);
-				}
-				if ((Unit > 0) && (Unit < 256))
-				{
-					self->Unit = Unit;
-				}
-				else
-				{
-					_log.Log(LOG_ERROR, "CPlugin:%s, illegal Unit number (%d), valid values range from 1 to 255.", __func__, Unit);
-					return 0;
-				}
-				if (TypeName) {
-					std::string	sTypeName = TypeName;
-
-					self->Type = pTypeGeneral;
-
-					if (sTypeName == "Pressure")					self->SubType = sTypePressure;
-					else if (sTypeName == "Percentage")				self->SubType = sTypePercentage;
-					else if (sTypeName == "Gas")
-					{
-						self->Type = pTypeP1Gas;
-						self->SubType = sTypeP1Gas;
-					}
-					else if (sTypeName == "Voltage")				self->SubType = sTypeVoltage;
-					else if (sTypeName == "Text")					self->SubType = sTypeTextStatus;
-					else if (sTypeName == "Switch")
-					{
-						self->Type = pTypeGeneralSwitch;
-						self->SubType = sSwitchGeneralSwitch;
-					}
-					else if (sTypeName == "Alert")
-					{
-						Py_DECREF(self->sValue);
-						self->sValue = PyUnicode_FromString("No Alert!");
-						self->SubType = sTypeAlert;
-					}
-					else if (sTypeName == "Current/Ampere")	
-					{
-						Py_DECREF(self->sValue);
-						self->sValue = PyUnicode_FromString("0.0;0.0;0.0");
-						self->Type = pTypeCURRENT;
-						self->SubType = sTypeELEC1;
-					}
-					else if (sTypeName == "Sound Level")			self->SubType = sTypeSoundLevel;
-					else if (sTypeName == "Barometer")
-					{
-						Py_DECREF(self->sValue);
-						self->sValue = PyUnicode_FromString("1021.34;0");
-						self->SubType = sTypeBaro;
-					}
-					else if (sTypeName == "Visibility")				self->SubType = sTypeVisibility;
-					else if (sTypeName == "Distance")				self->SubType = sTypeDistance;
-					else if (sTypeName == "Counter Incremental")	self->SubType = sTypeCounterIncremental;
-					else if (sTypeName == "Soil Moisture")			self->SubType = sTypeSoilMoisture;
-					else if (sTypeName == "Leaf Wetness")			self->SubType = sTypeLeafWetness;
-					else if (sTypeName == "kWh")
-					{
-						Py_DECREF(self->sValue);
-						self->sValue = PyUnicode_FromString("0; 0.0");
-						self->SubType = sTypeKwh;
-					}
-					else if (sTypeName == "Current (Single)")		self->SubType = sTypeCurrent;
-					else if (sTypeName == "Solar Radiation")		self->SubType = sTypeSolarRadiation;
-					else if (sTypeName == "Temperature")
-					{
-						self->Type = pTypeTEMP;
-						self->SubType = sTypeTEMP5;
-					}
-					else if (sTypeName == "Humidity")
-					{
-						self->Type = pTypeHUM;
-						self->SubType = sTypeHUM1;
-					}
-					else if (sTypeName == "Temp+Hum")
-					{
-						Py_DECREF(self->sValue);
-						self->sValue = PyUnicode_FromString("0.0;50;1");
-						self->Type = pTypeTEMP_HUM;
-						self->SubType = sTypeTH1;
-					}
-					else if (sTypeName == "Temp+Hum+Baro")
-					{
-						Py_DECREF(self->sValue);
-						self->sValue = PyUnicode_FromString("0.0;50;1;1010;1");
-						self->Type = pTypeTEMP_HUM_BARO;
-						self->SubType = sTypeTHB1;
-					}
-					else if (sTypeName == "Wind")
-					{
-						Py_DECREF(self->sValue);
-						self->sValue = PyUnicode_FromString("0;N;0;0;0;0");
-						self->Type = pTypeWIND;
-						self->SubType = sTypeWIND1;
-					}
-					else if (sTypeName == "Rain")
-					{
-						Py_DECREF(self->sValue);
-						self->sValue = PyUnicode_FromString("0;0");
-						self->Type = pTypeRAIN;
-						self->SubType = sTypeRAIN3;
-					}
-					else if (sTypeName == "UV")
-					{
-						Py_DECREF(self->sValue);
-						self->sValue = PyUnicode_FromString("0;0");
-						self->Type = pTypeUV;
-						self->SubType = sTypeUV1;
-					}
-					else if (sTypeName == "Air Quality")
-					{
-						self->Type = pTypeAirQuality;
-						self->SubType = sTypeVoltcraft;
-					}
-					else if (sTypeName == "Usage")
-					{
-						self->Type = pTypeUsage;
-						self->SubType = sTypeElectric;
-					}
-					else if (sTypeName == "Illumination")
-					{
-						self->Type = pTypeLux;
-						self->SubType = sTypeLux;
-					}
-					else if (sTypeName == "Waterflow")				self->SubType = sTypeWaterflow;
-					else if (sTypeName == "Wind+Temp+Chill")
-					{
-						Py_DECREF(self->sValue);
-						self->sValue = PyUnicode_FromString("0;N;0;0;0;0");
-						self->Type = pTypeWIND;
-						self->SubType = sTypeWIND4;
-					}
-					else if (sTypeName == "Selector Switch")
-					{
-						Py_DECREF(self->Options);
-						self->Options = PyUnicode_FromString("LevelActions:fHx8;LevelNames:T2ZmfExldmVsMXxMZXZlbDJ8TGV2ZWwz;LevelOffHidden:ZmFsc2U=;SelectorStyle:MA==");
-						self->Type = pTypeGeneralSwitch;
-						self->SubType = sSwitchTypeSelector;
-						self->SwitchType = 18;
-					}
-					else if (sTypeName == "Custom")					self->SubType = sTypeCustom;
-				}
-				if ((Type != -1) && Type) self->Type = Type;
-				if ((SubType != -1) && SubType) self->SubType = SubType;
-				if (SwitchType != -1) self->SwitchType = SwitchType;
-				if (Image != -1) self->Image = Image;
-				if (Options) {
-					Py_DECREF(self->Options);
-					self->Options = PyUnicode_FromString(Options);
-				}
-			}
-			else
-			{
-				CPlugin* pPlugin = NULL;
-				if (pModState) pPlugin = pModState->pPlugin;
-				_log.Log(LOG_ERROR, "Expected: myVar = Domoticz.Device(Name=\"myDevice\", Unit=0, TypeName=\"\", Type=0, Subtype=0, Switchtype=0, Image=0, Options=\"\")");
-				LogPythonException(pPlugin, __func__);
-			}
-		}
-		catch (std::exception e)
-		{
-			_log.Log(LOG_ERROR, "%s: Execption thrown: %s", __func__, e.what());
-		}
-		catch (...)
-		{
-			_log.Log(LOG_ERROR, "%s: Unknown execption thrown", __func__);
-		}
-
-		return 0;
-	}
-
-	static PyMemberDef CDevice_members[] = {
-		{ "ID",	T_INT, offsetof(CDevice, ID), READONLY, "Domoticz internal ID" },
-		{ "Name", T_OBJECT,	offsetof(CDevice, Name), READONLY, "Name" },
-		{ "nValue", T_INT, offsetof(CDevice, nValue), READONLY, "Numeric device value" },
-		{ "sValue", T_OBJECT, offsetof(CDevice, sValue), READONLY, "String device value" },
-		{ "SignalLevel", T_INT, offsetof(CDevice, SignalLevel), READONLY, "Numeric signal level" },
-		{ "BatteryLevel", T_INT, offsetof(CDevice, BatteryLevel), READONLY, "Numeric battery level" },
-		{ "Image", T_INT, offsetof(CDevice, Image), READONLY, "Numeric image number" },
-		{ "LastLevel", T_INT, offsetof(CDevice, LastLevel), READONLY, "Previous device level" },
-		{ "LastUpdate", T_OBJECT, offsetof(CDevice, LastUpdate), READONLY, "Last update timestamp" },
-		{ NULL }  /* Sentinel */
-	};
-
-	static PyObject* CDevice_refresh(CDevice* self)
-	{
-		if ((self->pPlugin) && (self->HwdID != -1) && (self->Unit != -1))
-		{
-			// load associated devices to make them available to python
-			std::vector<std::vector<std::string> > result;
-			result = m_sql.safe_query("SELECT Unit, ID, Name, nValue, sValue, DeviceID, Type, SubType, SwitchType, LastLevel, CustomImage, SignalLevel, BatteryLevel, LastUpdate FROM DeviceStatus WHERE (HardwareID==%d) AND (Unit==%d) ORDER BY Unit ASC", self->HwdID, self->Unit);
-			if (result.size() > 0)
-			{
-				for (std::vector<std::vector<std::string> >::const_iterator itt = result.begin(); itt != result.end(); ++itt)
-				{
-					std::vector<std::string> sd = *itt;
-					self->Unit = atoi(sd[0].c_str());
-					self->ID = atoi(sd[1].c_str());
-					Py_XDECREF(self->Name);
-					self->Name = PyUnicode_FromString(sd[2].c_str());
-					self->nValue = atoi(sd[3].c_str());
-					Py_XDECREF(self->sValue);
-					self->sValue = PyUnicode_FromString(sd[4].c_str());
-					Py_XDECREF(self->DeviceID);
-					self->DeviceID = PyUnicode_FromString(sd[5].c_str());
-					self->Type = atoi(sd[6].c_str());
-					self->SubType = atoi(sd[7].c_str());
-					self->SwitchType = atoi(sd[8].c_str());
-					self->LastLevel = atoi(sd[9].c_str());
-					self->Image = atoi(sd[10].c_str());
-					self->SignalLevel = atoi(sd[11].c_str());
-					self->BatteryLevel = atoi(sd[12].c_str());
-					Py_XDECREF(self->LastUpdate);
-					self->LastUpdate = PyUnicode_FromString(sd[13].c_str());
-				}
-			}
-		}
-		else
-		{
-			_log.Log(LOG_ERROR, "Device refresh failed, Device object is not associated with a plugin.");
-		}
-
-		Py_INCREF(Py_None);
-		return Py_None;
-	}
-
-	static PyObject* CDevice_insert(CDevice* self, PyObject *args)
-	{
-		if (self->pPlugin)
-		{
-			PyObject*	pNameBytes = PyUnicode_AsASCIIString(self->Name);
-			if (self->ID == -1)
-			{
-				if (self->pPlugin->m_bDebug)
-				{
-					_log.Log(LOG_NORM, "(%s) Creating device '%s'.", self->pPlugin->Name.c_str(), std::string(PyBytes_AsString(pNameBytes)).c_str());
-				}
-
-				std::vector<std::vector<std::string> > result;
-				result = m_sql.safe_query("SELECT Name FROM DeviceStatus WHERE (HardwareID==%d) AND (Unit==%d)", self->HwdID, self->Unit);
-				if (result.size() == 0)
-				{
-					char szID[40];
-					sprintf(szID, "%X%02X%02X%02X", 0, 0, (self->HwdID & 0xFF00) >> 8, self->HwdID & 0xFF);
-					PyObject*	pOptionBytes = PyUnicode_AsASCIIString(self->Options);
-					PyObject*	pSValueBytes = PyUnicode_AsASCIIString(self->sValue);
-					std::string	sLongName = self->pPlugin->Name + " - " + PyBytes_AsString(pNameBytes);
-					m_sql.safe_query(
-						"INSERT INTO DeviceStatus (HardwareID, DeviceID, Unit, Type, SubType, SwitchType, Used, SignalLevel, BatteryLevel, Name, nValue, sValue, CustomImage, Options) "
-						"VALUES (%d, '%q', %d, %d, %d, %d, 1, 12, 255, '%q', 0, '%q', %d, '%q')",
-						self->HwdID, szID, self->Unit, self->Type, self->SubType, self->SwitchType, sLongName.c_str(), std::string(PyBytes_AsString(pSValueBytes)).c_str(), self->Image, std::string(PyBytes_AsString(pOptionBytes)).c_str());
-					Py_DECREF(pOptionBytes);
-					Py_DECREF(pSValueBytes);
-
-					result = m_sql.safe_query("SELECT ID FROM DeviceStatus WHERE (HardwareID==%d) AND (Unit==%d)", self->HwdID, self->Unit);
-					if (result.size())
-					{
-						PyObject*	pKey = PyLong_FromLong(self->Unit);
-						if (PyDict_SetItem((PyObject*)self->pPlugin->m_DeviceDict, pKey, (PyObject*)self) == -1)
-						{
-							_log.Log(LOG_ERROR, "(%s) failed to add unit '%d' to device dictionary.", self->pPlugin->Name.c_str(), self->Unit);
-							Py_INCREF(Py_None);
-							return Py_None;
-						}
-
-						// Refresh device data to ensure it is usable straight away
-						PyObject* pRetVal = CDevice_refresh(self);
-						Py_DECREF(pRetVal);
-					}
-					else
-					{
-						_log.Log(LOG_ERROR, "(%s) Device creation failed, Hardware/Unit combination (%d:%d) not found in Domoticz.", self->pPlugin->Name.c_str(), self->HwdID, self->Unit);
-					}
-				}
-				else
-				{
-					_log.Log(LOG_ERROR, "(%s) Device creation failed, Hardware/Unit combination (%d:%d) already exists in Domoticz.", self->pPlugin->Name.c_str(), self->HwdID, self->Unit);
-				}
-			}
-			else
-			{
-				_log.Log(LOG_ERROR, "(%s) Device creation failed, '%s' already exists in Domoticz with Device ID '%d'.", self->pPlugin->Name.c_str(), std::string(PyBytes_AsString(pNameBytes)).c_str(), self->ID);
-			}
-			Py_DECREF(pNameBytes);
-		}
-		else
-		{
-			_log.Log(LOG_ERROR, "Device creation failed, Device object is not associated with a plugin.");
-		}
-
-		Py_INCREF(Py_None);
-		return Py_None;
-	}
-
-	static PyObject* CDevice_update(CDevice *self, PyObject *args, PyObject *kwds)
-	{
-		if (self->pPlugin)
-		{
-			int			nValue = self->nValue;
-			char*		sValue = NULL;
-			int			iSignalLevel = self->SignalLevel;
-			int			iBatteryLevel = self->BatteryLevel;
-			int			iImage = self->Image;
-			PyObject*	pNameBytes = PyUnicode_AsASCIIString(self->Name);
-			static char *kwlist[] = { "nValue", "sValue", "Image", "SignalLevel", "BatteryLevel", NULL };
-
-			if (!PyArg_ParseTupleAndKeywords(args, kwds, "is|iii", kwlist, &nValue, &sValue, &iImage, &iSignalLevel, &iBatteryLevel))
-			{
-				_log.Log(LOG_ERROR, "(%s) %s: Failed to parse parameters: 'nValue', 'sValue', 'SignalLevel' or 'BatteryLevel' expected.", __func__, PyBytes_AsString(pNameBytes));
-				LogPythonException(self->pPlugin, __func__);
-				Py_DECREF(pNameBytes);
-				return NULL;
-			}
-
-			if (self->pPlugin->m_bDebug)
-			{
-				wchar_t*	pWideValue = PyUnicode_AsWideCharString(self->sValue, NULL);
-				_log.Log(LOG_NORM, "(%s) Updating device from %d:'%S' to have values %d:'%s'.",
-					std::string(PyBytes_AsString(pNameBytes)).c_str(), self->nValue, pWideValue, nValue, sValue);
-				PyMem_Free(pWideValue);
-			}
-			PyObject*	pDeviceBytes = PyUnicode_AsASCIIString(self->DeviceID);
-			std::string	sName = PyBytes_AsString(pNameBytes);
-			m_sql.UpdateValue(self->HwdID, std::string(PyBytes_AsString(pDeviceBytes)).c_str(), (const unsigned char)self->Unit, (const unsigned char)self->Type, (const unsigned char)self->SubType, iSignalLevel, iBatteryLevel, nValue, std::string(sValue).c_str(), sName, true);
-			Py_DECREF(pDeviceBytes);
-			Py_DECREF(pNameBytes);
-
-			// Image change
-			if (iImage != self->Image)
-			{
-				time_t now = time(0);
-				struct tm ltime;
-				localtime_r(&now, &ltime);
-				m_sql.safe_query("UPDATE DeviceStatus SET CustomImage=%d, LastUpdate='%04d-%02d-%02d %02d:%02d:%02d' WHERE (HardwareID==%d) and (Unit==%d)",
-					iImage, ltime.tm_year + 1900, ltime.tm_mon + 1, ltime.tm_mday, ltime.tm_hour, ltime.tm_min, ltime.tm_sec, self->HwdID, self->Unit);
-			}
-
-			CDevice_refresh(self);
-		}
-		else
-		{
-			_log.Log(LOG_ERROR, "Device update failed, Device object is not associated with a plugin.");
-		}
-
-		Py_INCREF(Py_None);
-		return Py_None;
-	}
-
-	static PyObject* CDevice_delete(CDevice* self, PyObject *args)
-	{
-		if (self->pPlugin)
-		{
-			PyObject*	pNameBytes = PyUnicode_AsASCIIString(self->Name);
-			if (self->ID != -1)
-			{
-				if (self->pPlugin->m_bDebug)
-				{
-					_log.Log(LOG_NORM, "(%s) Deleting device '%s'.", self->pPlugin->Name.c_str(), std::string(PyBytes_AsString(pNameBytes)).c_str());
-				}
-
-				std::vector<std::vector<std::string> > result;
-				result = m_sql.safe_query("SELECT Name FROM DeviceStatus WHERE (HardwareID==%d) AND (Unit==%d)", self->HwdID, self->Unit);
-				if (result.size() != 0)
-				{
-					result = m_sql.safe_query("DELETE FROM DeviceStatus WHERE (HardwareID==%d) AND (Unit==%d)", self->HwdID, self->Unit);
-
-					PyObject*	pKey = PyLong_FromLong(self->Unit);
-					if (PyDict_DelItem((PyObject*)self->pPlugin->m_DeviceDict, pKey) == -1)
-					{
-						_log.Log(LOG_ERROR, "(%s) failed to delete unit '%d' from device dictionary.", self->pPlugin->Name.c_str(), self->Unit);
-						Py_INCREF(Py_None);
-						return Py_None;
-					}
-				}
-				else
-				{
-					_log.Log(LOG_ERROR, "(%s) Device deletion failed, Hardware/Unit combination (%d:%d) not found in Domoticz.", self->pPlugin->Name.c_str(), self->HwdID, self->Unit);
-				}
-			}
-			else
-			{
-				_log.Log(LOG_ERROR, "(%s) Device deletion failed, '%s' does not represent a device in Domoticz.", self->pPlugin->Name.c_str(), std::string(PyBytes_AsString(pNameBytes)).c_str());
-			}
-			Py_DECREF(pNameBytes);
-		}
-		else
-		{
-			_log.Log(LOG_ERROR, "Device deletion failed, Device object is not associated with a plugin.");
-		}
-
-		Py_INCREF(Py_None);
-		return Py_None;
-	}
-
-	static PyObject* CDevice_str(CDevice* self)
-	{
-		PyObject*	pRetVal = PyUnicode_FromFormat("ID: %d, Name: '%U', nValue: %d, sValue: '%U'", self->ID, self->Name, self->nValue, self->sValue);
-		return pRetVal;
-	}
-
-	static PyMethodDef CDevice_methods[] = {
-		{ "Refresh", (PyCFunction)CDevice_refresh, METH_NOARGS, "Refresh device details"},
-		{ "Create", (PyCFunction)CDevice_insert, METH_NOARGS, "Create the device in Domoticz." },
-		{ "Update", (PyCFunction)CDevice_update, METH_VARARGS | METH_KEYWORDS, "Update the device values in Domoticz." },
-		{ "Delete", (PyCFunction)CDevice_delete, METH_NOARGS, "Delete the device in Domoticz." },
-		{ NULL }  /* Sentinel */
-	};
-
-	static PyTypeObject CDeviceType = {
-		PyVarObject_HEAD_INIT(NULL, 0)
-		"Domoticz.Device",             /* tp_name */
-		sizeof(CDevice),             /* tp_basicsize */
-		0,                         /* tp_itemsize */
-		(destructor)CDevice_dealloc, /* tp_dealloc */
-		0,                         /* tp_print */
-		0,                         /* tp_getattr */
-		0,                         /* tp_setattr */
-		0,                         /* tp_reserved */
-		0,                         /* tp_repr */
-		0,                         /* tp_as_number */
-		0,                         /* tp_as_sequence */
-		0,                         /* tp_as_mapping */
-		0,                         /* tp_hash  */
-		0,                         /* tp_call */
-		(reprfunc)CDevice_str,     /* tp_str */
-		0,                         /* tp_getattro */
-		0,                         /* tp_setattro */
-		0,                         /* tp_as_buffer */
-		Py_TPFLAGS_DEFAULT |
-		Py_TPFLAGS_BASETYPE,   /* tp_flags */
-		"Domoticz Device",           /* tp_doc */
-		0,                         /* tp_traverse */
-		0,                         /* tp_clear */
-		0,                         /* tp_richcompare */
-		0,                         /* tp_weaklistoffset */
-		0,                         /* tp_iter */
-		0,                         /* tp_iternext */
-		CDevice_methods,             /* tp_methods */
-		CDevice_members,             /* tp_members */
-		0,                         /* tp_getset */
-		0,                         /* tp_base */
-		0,                         /* tp_dict */
-		0,                         /* tp_descr_get */
-		0,                         /* tp_descr_set */
-		0,                         /* tp_dictoffset */
-		(initproc)CDevice_init,      /* tp_init */
-		0,                         /* tp_alloc */
-		CDevice_new                 /* tp_new */
-	};
-
-	static PyModuleDef CDevice_module = {
-		PyModuleDef_HEAD_INIT,
-		"Device",
-		"Domoticz Device",
-		-1,
-		NULL, NULL, NULL, NULL, NULL
 	};
 
 	PyMODINIT_FUNC PyInit_Domoticz(void)
@@ -1592,7 +1003,7 @@ namespace Plugins {
 			break;
 		case PMT_Command:
 			sHandler = "onCommand";
-			if (Message.m_fLevel != -273.15)
+			if (Message.m_fLevel != -273.15f)
 			{
 				pParams = Py_BuildValue("isfi", Message.m_Unit, Message.m_Message.c_str(), Message.m_fLevel, 0);
 			}
@@ -1877,6 +1288,7 @@ namespace Plugins {
 			PluginMessageQueue.push(Message);
 		}
 	}
+
 	void CPlugin::SendCommand(const int Unit, const std::string & command, const float level)
 	{
 		//	Add command to message queue
