+++ msbuild/domoticz.vcxproj.filters
@@ -1981,6 +1981,9 @@
     <ClInclude Include="..\push\BasePush.h">
       <Filter>Pushers</Filter>
     </ClInclude>
+    <ClInclude Include="..\hardware\plugins\PythonObjects.h">
+      <Filter>Devices\Python Plugins</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\webserver\Base64.cpp">
@@ -2624,6 +2627,9 @@
     </ClCompile>
     <ClCompile Include="..\push\BasePush.cpp">
       <Filter>Pushers</Filter>
+    </ClCompile>
+    <ClCompile Include="..\hardware\plugins\PythonObjects.cpp">
+      <Filter>Devices\Python Plugins</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
